refactor(db): move some method from `user.go` to `users.go`
