+++ internal/db/repo.go
@@ -1157,8 +1157,8 @@ func (err ErrReachLimitOfRepo) Error() string {
 
 // CreateRepository creates a repository for given user or organization.
 func CreateRepository(doer, owner *User, opts CreateRepoOptionsLegacy) (_ *Repository, err error) {
-	if !owner.CanCreateRepo() {
-		return nil, ErrReachLimitOfRepo{Limit: owner.RepoCreationNum()}
+	if !owner.canCreateRepo() {
+		return nil, ErrReachLimitOfRepo{Limit: owner.MaxNumRepos()}
 	}
 
 	repo := &Repository{
@@ -2467,8 +2467,8 @@ func HasForkedRepo(ownerID, repoID int64) (*Repository, bool, error) {
 
 // ForkRepository creates a fork of target repository under another user domain.
 func ForkRepository(doer, owner *User, baseRepo *Repository, name, desc string) (_ *Repository, err error) {
-	if !owner.CanCreateRepo() {
-		return nil, ErrReachLimitOfRepo{Limit: owner.RepoCreationNum()}
+	if !owner.canCreateRepo() {
+		return nil, ErrReachLimitOfRepo{Limit: owner.MaxNumRepos()}
 	}
 
 	repo := &Repository{
