+++ internal/route/api/v1/user/user.go
@@ -19,7 +19,7 @@ import (
 func Search(c *context.APIContext) {
 	opts := &db.SearchUserOptions{
 		Keyword:  c.Query("q"),
-		Type:     db.UserIndividual,
+		Type:     db.UserTypeIndividual,
 		PageSize: com.StrTo(c.Query("limit")).MustInt(),
 	}
 	if opts.PageSize == 0 {
